FoodRecipeApi/requirements.txt
Django==4.0.4
djangorestframework==3.13.1
psycopg2==2.9.3
Pillow==9.1.0

flake>=3.6.0,<3.7.0
